proptypes basic types validators moved to one file
