--- react/src/proptypes/util/IsBoolean.js
@@ -1,3 +1,0 @@
-const isBoolean = value => value && typeof value === "boolean";
-
-export default isBoolean;
