--- react/src/proptypes/util/IsNumber.js
@@ -1,3 +1,0 @@
-const isNumber = value => value && typeof value === "number";
-
-export default isNumber;
