--- react/src/proptypes/util/IsString.js
@@ -1,3 +1,0 @@
-const isString = value => value && typeof value === "string";
-
-export default isString;
