--- react/src/proptypes/util/IsTime.js
@@ -1,5 +1,0 @@
-import { isNumber } from "util";
-
-const isTime = value => value && isNumber(value.hour) && isNumber(value.minute);
-
-export default isTime;
